Slight reworking of the structure of the alignment algorithm in bwt_algorithms.cpp

+++ src/SuffixTools/BWT.cpp
@@ -10,6 +10,7 @@
 #include "Timer.h"
 #include <istream>
 #include <queue>
+#include <inttypes.h>
 
 struct PartialAlign
 {
@@ -96,8 +97,8 @@ void BWT::backwardSearch(std::string w) const
 		curr = w[j];
 		printf("RL = C(%c) + O(%c,%d) + %zu\n", curr, curr, r_lower - 1, m_numStrings); 
 		printf("RU = C(%c) + O(%c,%d)\n", curr, curr, r_upper); 
-		printf("RL = %llu + %llu + %zu\n", PRED(curr), OCC(curr, r_lower - 1), m_numStrings); 
-		printf("RU = %llu + %llu\n", PRED(curr), OCC(curr, r_upper)); 
+		printf("RL = %zu + %zu + %zu\n", (size_t)PRED(curr), (size_t)OCC(curr, r_lower - 1), m_numStrings); 
+		printf("RU = %zu + %zu\n", (size_t)PRED(curr), (size_t)OCC(curr, r_upper)); 
 		r_lower = PRED(curr) + OCC(curr, r_lower - 1);
 		r_upper = PRED(curr) + OCC(curr, r_upper) - 1;
 		printf("Curr: %c, Interval now: %d,%d\n", curr, r_lower, r_upper);
